split out AppFrame and SideNav
